+++ skorch/net.py
@@ -1345,7 +1345,10 @@ class NeuralNet:
         for key in kwargs:
             if key.endswith('_'):
                 continue
-            for prefix in self.prefixes_:
+
+            # see https://github.com/skorch-dev/skorch/pull/590 for
+            # why this must be sorted
+            for prefix in sorted(self.prefixes_, key=lambda s: (-len(s), s)):
                 if key.startswith(prefix):
                     if not key.startswith(prefix + '__'):
                         missing_dunder_kwargs.append((prefix, key))